moving MultiObject traits to traits_extension
